# swagger/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "API Documentation",
    "version": "1.0.0",
    "description": "Documentation de l'API hackapi"
  },
  "servers": [
    {
      "url": "http://localhost:4000",
      "description": "Serveur local"
    }
  ],
  "paths": {
...
    "/features/phishing-service": {
      "get": {
        "summary": "Scraper un site et générer un fichier HTML modifié",
        "tags": [
          "Feature - Phishing service"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "url",
            "required": true,
            "description": "URL du site à scraper",
            "schema": {
              "type": "string",
              "format": "uri"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Fichier HTML généré avec succès",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "description": "Message de succès avec l'emplacement du fichier généré"
                    },
                    "info": {
                      "type": "string",
                      "description": "L'URL pour récupérer les données soumises"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "L'URL est requise",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Erreur lors du scraping du site",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Soumettre les données d'un formulaire",
        "tags": [
          "Feature - Phishing service"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "site": {
                    "type": "string",
                    "description": "L'URL du site"
                  },
                  "[ autres champs du formulaire ]": {
                    "type": "string",
                    "description": "Autres données soumises par l'utilisateur"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Données reçues avec succès",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "description": "Message de succès"
                    },
                    "submission": {
                      "type": "object",
                      "description": "Données soumises"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Aucune donnée valide trouvée dans le formulaire",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Erreur lors de la soumission des données",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/features/phishing-service/datas": {
      "get": {
        "summary": "Récupérer toutes les soumissions de formulaires",
        "tags": [
          "Feature - Phishing service"
        ],
        "responses": {
          "200": {
            "description": "Liste des soumissions récupérées avec succès",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "site": {
                        "type": "string",
                        "description": "L'URL du site"
                      },
                      "[ autres champs du formulaire ]": {
                        "type": "string",
                        "description": "Autres données soumises"
                      },
                      "timestamp": {
                        "type": "string",
                        "format": "date-time",
                        "description": "Horodatage de la soumission"
                      }
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Aucune soumission trouvée",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Erreur lors de la récupération des données",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/features/retrieve-all-domains": {
      "get": {
        "summary": "Récupère tous les sous-domaines d'un domaine donné",
        "tags": [
          "Feature - Retrieve all domains & subdomains"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "domain",
            "required": true,
            "description": "Le nom de domaine pour lequel récupérer les sous-domaines",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Sous-domaines récupérés avec succès",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "domain": {
                      "type": "string",
                      "description": "Le domaine principal"
                    },
                    "subdomains": {
                      "type": "array",
                      "items": {
                        "type": "string"
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad Request - Domaine manquant",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
...
}